refactor(landing): move landing composition to components/landing
app/page.tsx previously imported the landing composition from
pages/landing/index.tsx, coupling the App Router entry route to the
legacy Pages Router. Relocate that composition to
components/landing/landing-page.tsx (alongside the rest of
components/landing/*) and point app/page.tsx at the new module so the
`/` route no longer depends on pages/.
