+++ app/page.tsx
@@ -1,4 +1,4 @@
-import LandingPage from "@/pages/landing";
+import LandingPage from "@/components/landing/landing-page";
 
 export default function Home() {
   return (
